Fixed TBB_Timeseries and TBB_Station group to respect dipole selection when reading out attributes.
The `getAttributes` method now returns data only for the set of selected dipoles.
Also added a `nofSelectedDipoles` method to get the total number of dipoles in a dataset that have been
selected.
Finally fixed a bug that caused the DATA_LENGTH attribute to be read for the first dipole only.