Include fancyhdr in LaTeX header.
In later versions of Doxygen (at least 1.7.4, with which I'm testing), this
seems to be automatically added to doxygen.sty (presumably based on its
inclusion in the EXTRA_PACKAGES directive in the Doxyfile).

In older versions of Doxygen (including 1.5.5, as used on LOFAR CEP1), the
EXTRA_PACKAGES are rather used in the generation of the LaTeX header. However,
the DAL eschews this header in favour of its own. fancyhdr.sty is therefore
not loaded, and the documentation build fails.